backend/requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
uvicorn[standard]
pydantic-settings
python-dotenv
pandas
numpy
requests
